Initialize system clock using network time protocol
These changes add an initialization step which synchronizes the system
clock by polling the current time from a network server. This is
necessary because the flight computer does not maintain its internal
clock when it is powered off.

+++ visualizer/src/main.c
@@ -8,7 +8,7 @@
 
 int main(int argc, char *argv[argc+1]) {
   setup_serial_stream("/dev/cu.usbserial-0001");
-  uint8_t readBuffer[128];
+  uint64_t readBuffer[128];
   int readCount = 0;
   while(readCount <= 250) {
     int numberOfBytesRead = read_serial_stream(readBuffer, sizeof(readBuffer));
